# Improve `HTML.escape(string, io)` performance