--- a/src/html.cr
+++ b/src/html.cr
@@ -58,7 +58,7 @@ module HTML
             end
       if str
         io.write_string(string[last_copy_at, index &- last_copy_at])
-        last_copy_at = index
+        last_copy_at = index &+ 1
         io << str
       end
     end