This code was originally created because I needed to escape a whole file to HTML and didn't want the overhead from copying everything to a string.

There's a new method for escaping from `Bytes`, and the method for escaping from `String` now uses this implementation.

The new implementation iterates over the individual bytes rather than each char. But because all escapable chars are one byte long (and because of the way UTF-8 works), this is not a problem.

<details>
<summary>Benchmark Code</summary>

```cr
require "html"
require "benchmark"

module Test
  @[AlwaysInline]
  def self.escape(string : String, io : IO) : Nil
    escape(string.to_slice, io)
  end

  def self.escape(string : Bytes, io : IO) : Nil
    last_copy_at : Int32 = 0
    string.each_with_index do |byte, index|
      str = case byte
            when '&'  then "&amp;"
            when '<'  then "&lt;"
            when '>'  then "&gt;"
            when '"'  then "&quot;"
            when '\'' then "&#39;"
            else
              nil
            end
      if str
        io.write_string(string[last_copy_at, index &- last_copy_at])
        last_copy_at = index &+ 1
        io << str
      end
    end
    io.write_string(string[last_copy_at, string.size &- last_copy_at])
  end
end

io = IO::Memory.new
test1 = "This is a string without escapable chars"
test2 = test1 * 1000
test3 = "Ä string withööt escäpäbles, büt with ünicöde chärs"
test4 = test3 * 1000
test5 = "ASCII-Only \"string\" <with> &escapables&"
test6 = test5 * 1000
test7 = "Ünicode \"with\" <<escäpäbles>>"
test8 = test7 * 1000

{% for var in %w(test1 test2 test3 test4 test5 test6 test7 test8) %}
  Benchmark.ips do |x|
    x.report("old") { HTML.escape({{var.id}}, io.rewind) }
    x.report("new") { Test.escape({{var.id}}, io.rewind) }
  end
  puts "---"
{% end %}
```
</details>

<details>
<summary>Benchmark results</summary>

```
old   2.91M (343.91ns) (± 3.09%)  0.0B/op  11.86× slower
new  34.47M ( 29.01ns) (± 2.49%)  0.0B/op        fastest
---
old   3.04k (328.68µs) (± 1.57%)  0.0B/op  15.89× slower
new  48.36k ( 20.68µs) (± 1.21%)  0.0B/op        fastest
---
old   1.70M (588.47ns) (± 1.72%)  0.0B/op  14.33× slower
new  24.35M ( 41.07ns) (± 3.61%)  0.0B/op        fastest
---
old   1.71k (585.42µs) (± 1.49%)  0.0B/op  13.78× slower
new  23.54k ( 42.47µs) (± 1.12%)  0.0B/op        fastest
---
old   2.61M (382.77ns) (± 3.16%)  0.0B/op   3.60× slower
new   9.40M (106.37ns) (± 3.45%)  0.0B/op        fastest
---
old   2.75k (363.96µs) (± 0.76%)  0.0B/op   3.70× slower
new  10.16k ( 98.40µs) (± 0.64%)  0.0B/op        fastest
---
old   3.10M (323.03ns) (± 0.47%)  0.0B/op   3.32× slower
new  10.27M ( 97.35ns) (± 1.71%)  0.0B/op        fastest
---
old   3.13k (319.25µs) (± 0.37%)  0.0B/op   3.33× slower
new  10.44k ( 95.83µs) (± 0.81%)  0.0B/op        fastest
---
```
</details>